postNewBooking to json
