+++ package.json
@@ -9,6 +9,7 @@
     "@mui/material": "^5.10.9",
     "@mui/styled-engine-sc": "^5.10.6",
     "@mui/x-date-pickers": "^5.0.4",
+    "axios": "^1.1.3",
     "dayjs": "^1.11.5",
     "react": "^17.0.2",
     "react-app-provider": "^1.0.6",
