+++ src/App.js
@@ -13,18 +13,18 @@ import { Toaster } from "react-hot-toast"
 
 function App () {
   return (
-        <Router>
-        <div><Toaster/></div>
-          <Switch>
-            <Route path ={routes.loginpage} component={loginpage}/>
-            <Route exact path = {routes.home} component={Home}/>
-            <Route path = {routes.userDetailsPage} component={UserAccount}/>
-            <Route path = {routes.bookingPage} component={BookingPage}/>
-            <Route path = {routes.predictionPage} component={PredictionPage}/>
-            <Route path = {routes.scoreBoardPage} component={ScoreBoard}/>
-            <Route path = {routes.test} component={TestView}/>
-          </Switch>
-        </Router>
+    <Router>
+      <div><Toaster/></div>
+      <Switch>
+        <Route path ={routes.loginpage} component={loginpage}/>
+        <Route exact path = {routes.home} component={Home}/>
+        <Route path = {routes.userDetailsPage} component={UserAccount}/>
+        <Route path = {routes.bookingPage} component={BookingPage}/>
+        <Route path = {routes.predictionPage} component={PredictionPage}/>
+        <Route path = {routes.scoreBoardPage} component={ScoreBoard}/>
+        <Route path = {routes.test} component={TestView}/>
+      </Switch>
+    </Router>
   )
 }
