+++ src/components/util/util.js
@@ -1,3 +1,6 @@
+import axios from "axios"
+
+// const BASE_URL = "http://localhost:8000"
 
 export const routes = {
   "loginpage": "/login",
@@ -9,4 +12,8 @@ export const routes = {
   "predictionPage": "/prediction-page",
   "scoreBoardPage": "/scoreboard-page",
   "test": "/test"
+}
+
+export const postNewBooking = async (jsonValue) => {
+  await axios.post("http://localhost:8000/bookings", { jsonValue })
 }
