+++ src/index.css
@@ -109,7 +109,7 @@ Button:hover{
 
 } */
 .dateTime{
-  margin-left: 200px;
-  margin-right:200px;
+  margin-left: 400px;
+  margin-right:400px;
   padding-bottom: 220px;
 }
